refractor: rename users directory to user
