+++ app/src/main/java/com/github/onlynotesswent/model/user/UserRepositoryFirestore.kt
@@ -1,4 +1,4 @@
-package com.github.onlynotesswent.model.users
+package com.github.onlynotesswent.model.user
 
 import com.google.firebase.Timestamp
 import com.google.firebase.auth.FirebaseAuth
