+++ app/src/main/java/com/github/onlynotesswent/ui/overview/EditMarkdownScreen.kt
@@ -53,7 +53,7 @@ import androidx.compose.ui.unit.dp
 import com.github.onlynotesswent.model.file.FileType
 import com.github.onlynotesswent.model.file.FileViewModel
 import com.github.onlynotesswent.model.note.NoteViewModel
-import com.github.onlynotesswent.model.users.UserViewModel
+import com.github.onlynotesswent.model.user.UserViewModel
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.utils.ScreenTopBar
 import com.mohamedrejeb.richeditor.model.RichTextState
